+++ cmd/nba-api-server/main.go
@@ -16,7 +16,7 @@ import (
 	"github.com/n-ae/nba-api-go/v3/pkg/stats"
 )
 
-const version = "3.1.2"
+const version = "3.1.3"
 
 var (
 	buildTime = "unknown"